Move getter functions that depend on plugins
