+++ src/app/state/coordination.js
@@ -1,10 +1,4 @@
-import { fromEntries } from '../../utils';
 import { CoordinationType, Component } from '../constants';
-import {
-  getPluginCoordinationTypeDefaults,
-  getPluginViewTypes,
-  getPluginCoordinationTypesForViewType,
-} from '../plugins';
 
 /**
  * Coordination types may have default values,
@@ -62,15 +56,6 @@ export const DEFAULT_COORDINATION_VALUES = {
   [CoordinationType.ADDITIONAL_CELL_SETS]: null,
   [CoordinationType.MOLECULE_HIGHLIGHT]: null,
 };
-
-// Need to do this in a function since the plugin coordination
-// types are dynamic.
-export function getDefaultCoordinationValues() {
-  return {
-    ...DEFAULT_COORDINATION_VALUES,
-    ...getPluginCoordinationTypeDefaults(),
-  };
-}
 
 // The following coordination types should be
 // initialized to independent scopes when
@@ -268,15 +253,3 @@ export const COMPONENT_COORDINATION_TYPES = {
     CoordinationType.GENE_SELECTION,
   ],
 };
-
-// Need to do this in a function since the plugin coordination
-// types are dynamic.
-export function getComponentCoordinationTypes() {
-  return {
-    ...COMPONENT_COORDINATION_TYPES,
-    ...fromEntries(getPluginViewTypes().map(viewType => ([
-      viewType,
-      getPluginCoordinationTypesForViewType(viewType),
-    ]))),
-  };
-}
